+++ app/frontend/src/javascript/components/dashboard/reservations/reservations-panel.tsx
@@ -8,6 +8,8 @@ import { Loader } from '../../base/loader';
 import FormatLib from '../../../lib/format';
 import { FabPopover } from '../../base/fab-popover';
 import { useImmer } from 'use-immer';
+import _ from 'lodash';
+import { FabButton } from '../../base/fab-button';
 
 interface SpaceReservationsProps {
   userId: number,
@@ -23,6 +25,7 @@ const ReservationsPanel: React.FC<SpaceReservationsProps> = ({ userId, onError, 
 
   const [reservations, setReservations] = useState<Array<Reservation>>([]);
   const [details, updateDetails] = useImmer<Record<number, boolean>>({});
+  const [showMore, setShowMore] = useState<boolean>(false);
 
   useEffect(() => {
     ReservationAPI.index({ user_id: userId, reservable_type: reservableType })
@@ -70,12 +73,30 @@ const ReservationsPanel: React.FC<SpaceReservationsProps> = ({ userId, onError, 
   };
 
   /**
+   * Shows/hide the very old reservations list
+   */
+  const toggleShowMore = (): void => {
+    setShowMore(!showMore);
+  };
+
+  /**
+   * Display a placeholder when there's no reservation to display
+   */
+  const noReservations = (): ReactNode => {
+    return (
+      <li className="no-reservations">{t('app.logged.dashboard.reservations.reservations_panel.no_reservations')}</li>
+    );
+  };
+
+  /**
    * Render the reservation in a user-friendly way
    */
   const renderReservation = (reservation: Reservation, state: 'passed' | 'futur'): ReactNode => {
     return (
       <li key={reservation.id} className="reservation">
-        <a className={`reservation-title ${details[reservation.id] ? 'clicked' : ''}`} onClick={toggleDetails(reservation.id)}>{reservation.reservable.name}</a>
+        <a className={`reservation-title ${details[reservation.id] ? 'clicked' : ''}`} onClick={toggleDetails(reservation.id)}>
+          {reservation.reservable.name} - {FormatLib.date(reservation.slots_attributes[0].start_at)}
+        </a>
         {details[reservation.id] && <FabPopover title={t('app.logged.dashboard.reservations.reservations_panel.slots_details')}>
           {reservation.slots_attributes.filter(s => filterSlot(s, state)).map(
             slot => <span key={slot.id} className="slot-details">
@@ -87,15 +108,24 @@ const ReservationsPanel: React.FC<SpaceReservationsProps> = ({ userId, onError, 
     );
   };
 
+  const futur = reservationsByDate('futur');
+  const passed = _.orderBy(reservationsByDate('passed'), r => r.slots_attributes[0].start_at, 'desc');
+
   return (
     <FabPanel className="reservations-panel" header={header()}>
       <h4>{t('app.logged.dashboard.reservations.reservations_panel.upcoming')}</h4>
       <ul>
-        {reservationsByDate('futur').map(r => renderReservation(r, 'futur'))}
+        {futur.length === 0 && noReservations()}
+        {futur.map(r => renderReservation(r, 'futur'))}
       </ul>
       <h4>{t('app.logged.dashboard.reservations.reservations_panel.passed')}</h4>
       <ul>
-      {reservationsByDate('passed').reverse().map(r => renderReservation(r, 'passed'))}
+        {passed.length === 0 && noReservations()}
+        {passed.slice(0, 10).map(r => renderReservation(r, 'passed'))}
+        {passed.length > 10 && !showMore && <li className="show-more"><FabButton onClick={toggleShowMore}>
+          {t('app.logged.dashboard.reservations.reservations_panel.show_more')}
+        </FabButton></li>}
+        {passed.length > 10 && showMore && passed.slice(10).map(r => renderReservation(r, 'passed'))}
       </ul>
     </FabPanel>
   );